reform search widget
